+++ amaranth/rpc.py
@@ -3,7 +3,7 @@ import json
 import argparse
 import importlib
 
-from .hdl import Signal, Record, Elaboratable
+from .hdl import Signal, Elaboratable
 from .back import rtlil
 
 
@@ -69,10 +69,9 @@ def _serve_yosys(modules):
             try:
                 elaboratable = modules[module_name](*args, **kwargs)
                 ports = []
-                # By convention, any public attribute that is a Signal or a Record is
-                # considered a port.
+                # By convention, any public attribute that is a Signal is considered a port.
                 for port_name, port in vars(elaboratable).items():
-                    if not port_name.startswith("_") and isinstance(port, (Signal, Record)):
+                    if not port_name.startswith("_") and isinstance(port, Signal):
                         ports += port._lhs_signals()
                 rtlil_text = rtlil.convert(elaboratable, name=module_name, ports=ports)
                 response = {"frontend": "ilang", "source": rtlil_text}